[CORE] Move all Abstracts to the Interfaces and see if this implementation for ArmorStandManipulationEvent works?
